fix: QS footer on mobile version
